Make Coordinates a newtype instead of a type
This is necessary to define a custom Ord instance for Coordinates:

We want to order coordinates based on their distance to the center
of the playing field instead of ordering on x first and on y second.
This ordering will allow us to effectively have an indexed version
of the playing field without changing its current representation.